requirements.txt
openpyxl>=3.1.0
pyyaml>=6.0
rich>=13.0.0
openai>=1.0.0
PyPDF2>=3.0.0

